fix the master branch, squashing following commits  e008389ad NPM and venice integ test fixes (#9472)  f477bd90b Fwlogs query (#9024)  cb6f7f9cc DelphiGo: Print error if reactor is hogging the go routine (#9435)  bd5662cd7 ionic: leave macvlan offload off by default  d53b1a64f ionic: add tx checksum stats  7f8cf7b6c ionic: fix tx wake-drop numbers  60cb468cb pciemgr: coverity issues addressed  e6600761d pciemgr: debug bar supports prefetch enable/disable  bbeb31fd1 Merge pull request #9436 from pavramas/ENABLE_TELEMETRY  f7b9aebe6 Adding changes for RC_AGAIN  852d3ccc2 Merge pull request #9461 from pirabhur/master  82e4eb1f4 moved to sync.Map() to fix concurrent access (#9463)  81707e909 Handle failures gracefully in events recorder (#9446)  83933edd6 Merge pull request #9385 from akoradha/ak_pb  1b11f2fac Merge pull request #9457 from hareeshrc/nic_asic3  a21aba9b2 Display IP Address in the Status of penctl show mode command (#9450)  026ff32e4 Restore minimal workspace file and gtest build artifact to keep storage jobd tests functional  3a0bf9050 upgrade manager goes upto switch root (#9454)  b4d9ccd24 capview fix + add asic files to assets  f1a2ec9cd ionic_rdma: fixes for issues found by coverity (#9414)  60d256459 Files required for capview  6f3901cd5 Fix the TestServerStartStop (#9447)  b21869d0e Fix iota sanity (#9451)  169cb1f42 Merge pull request #9067 from pavramas/IOTA_ALG  65b526609 Merge pull request #9434 from pavramas/FIX_TIMER  8b9d79591 Add App.json to IOTA  848232194 Remove timer deletes from session timer callbacks to avoid deadlock. Let the timer library handle the deletes. Fixing another deadlock to not get config db lock for session get_all and delete_all as FTE individually gets it anyway  5f14242e4 Remove Ignore flag from telemetry suite  93d78c8e1 Removing nic/asic copy to third-party  24b94028c QOS class delete + fix dscp cos programming

+++ nic/sdk/platform/pciemgr/src/indirect.c
@@ -85,9 +85,9 @@ aximst_addr(const unsigned int port,
     assert(entry < AXIMST_ENTRIES_PER_PORT);
 
     return (AXIMST_BASE + 
-            (idx * AXIMST_STRIDE) +
-            (port * AXIMST_PORT_STRIDE) +
-            (entry * AXIMST_ENTRY_STRIDE));
+            ((u_int64_t)idx * AXIMST_STRIDE) +
+            ((u_int64_t)port * AXIMST_PORT_STRIDE) +
+            ((u_int64_t)entry * AXIMST_ENTRY_STRIDE));
 }
 
 static void
